==> INDI-genetics/scRNA-seq analysis/iPSC/01_call_cell.Rmd <==
---
title: "hiPSCi cell line evaluation (iPSC) - 01 Cell Calling"
output: html_notebook
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.width=6, fig.height=4, fig.path='Figs/', echo=FALSE, warning=FALSE, message=FALSE)
```

```{r, warning=FALSE}
# Load required packages
library(DropletUtils)
library(Matrix)
library(scater) 
library(scran)
library(BiocSingular)
library(BiocParallel)
library(pheatmap)
library(plotly)
library(batchelor)
library(viridis)
library(reshape2)
library(SingleCellExperiment)
library(cowplot)
ncores = 3
mcparam = MulticoreParam(workers = ncores)
register(mcparam)
```

For this experiment, we have 1 sample composed of all 8 iPSC cell line (undifferentiated) which have been multiplexed by genotype. 

# Cell calling

We will use the `EmptyDrops` method for performing cell selection, developed by Aaron, since the implementation by CellRanger is not exactly the same as the implementation in the R package.

```{r}
samples = list.files(pattern="Sample_.*")

sce_object = lapply(samples, read10xCounts)
```

Let's first look at the barcode distribution across cell, ranked by total UMI count.

```{r, echo=FALSE, message=FALSE, warning=FALSE, results=FALSE}
lapply(sce_object, function(x){
  bcrank = barcodeRanks(counts(x))
  
  # Only showing unique points for plotting speed.
  uniq = !duplicated(bcrank$rank)
  plot(bcrank$rank[uniq], bcrank$total[uniq], log="xy", xlab="Rank", ylab="Total UMI count", cex.lab=1.2)
  
  abline(h=metadata(bcrank)$inflection, col="darkgreen", lty=2)
  abline(h=metadata(bcrank)$knee, col="dodgerblue", lty=2)
  
  legend("bottomleft", legend=c("Inflection", "Knee"),col=c("darkgreen", "dodgerblue"), lty=2, cex=1.2)
})
```

`EmptyDrops` by default uses any barcode with < 100 UMI count to model the ambient RNA content as these barcodes are assumed to correspond to empty droplets. Let's look at these background RNA distribution.

```{r echo=FALSE, message=FALSE, warning=FALSE, results=FALSE}
lapply(sce_object, function(x){
  mat = assay(x)
  targets = mat[, Matrix::colSums(mat)!=0 & Matrix::colSums(mat) < 100]
  sums = Matrix::colSums(targets)
  reads = as.matrix(table(sums) * as.numeric(names(table(sums))))
  
  ggplot(data.frame(lib = reads, n = as.numeric(rownames(reads))), aes(x = n, y= lib)) +
    geom_bar(stat = "identity") +
    theme_bw() +
    labs(x = "Barcode library UMI count", y = "Total contribution to background vector") +
    scale_x_continuous(breaks = seq(0, 100, 5))
})
```

For the iPSC sample, 100 is an acceptable threshold to use.

Let's perform `EmptyDrops` now.

```{r echo=FALSE, message=FALSE, warning=FALSE, results=FALSE}
set.seed(100)
lower_bound = 100
emptyDrops_out = lapply(sce_object, function(x) {
  emptyDrops(counts(x), lower = lower_bound, BPPARAM = mcparam, niters = 10000)
})
```

```{r}
lapply(emptyDrops_out, function(x) summary(x$FDR <= 0.001))
```

```{r}
sce_object = lapply(1:length(sce_object), function (i) {
  print(i)
  sce_object[[i]][,which(emptyDrops_out[[i]]$FDR <= 0.001)]
})
```

Valid barcode extraction for vartrix-vireo cell line identification.

```{r}
lapply(1:length(sce_object), function (i) {
  write.table(sce_object[[i]]$Barcode, file = paste("emptyDrops_barcode_", sce_object[[i]]$Sample[1], "_vartrix.txt", sep=""), quote=FALSE, row.names = FALSE, col.names = FALSE)
})
```


```{r}
saveRDS(emptyDrops_out, "emptyDrops_output.rds")
saveRDS(sce_object, "called_cell.rds")
```
